Changed architecture of the project to follow the UML component diagram closely: components are now more decoupled from each-other, CLI is split up into several files, there exists a converter manager which picks which conversion type to use, unoserver converter is now in its own file
